Create Love Kush
Love Kush (100 Marks)
Love and Kush are brothers studying in the same school. Love is elder to Kush. They both got into an argument about who is more famous in the school. To decide who is more famous, they thought of conducting elections. They asked the students to write the initials of the person who they think is more famous.


Love is over-confident and says to Kush that he will only win if he has more than twice the number of votes of Kush strictly. They both trust you to be fair and want the results to be declared by you. Please declare the results of the elections.




Input Format
The first line of input consists of a number of test cases, T.

Next T lines each consist of a string of lowercase letters, l and k only.

Constraints
1<= T <=10

1<= |string_length| <=1000



Output Format
Print the winner of the election. Print Love if Love wins the election otherwise print Kush.

Sample TestCase 1
Input
2
llkklll
kkkkll
Output
Love
Kush
Explanation

Test Case 1:

Number of votes for Love = 5

Number of votes for Kush = 2


Since, the number of votes for Love are more than twice of Kush, Love wins the election.


Test Case 2:

Number of votes for Love = 2

Number of votes for Kush = 4


Since, the number of votes for Kush are more than Love, Kush wins the election. This solution reads in the number of test cases and then loops through each test case. For each test case, it reads in the vote string and counts the number of votes for Love and Kush. It then checks if Love has more than twice the number of votes as Kush, and if so, prints "Love" as the winner. Otherwise, it prints "Kush" as the winner.